Add Common class library
Move swagger to common
