+++ Apps/Common/Swagger/ServicesExtensions.cs
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 //-------------------------------------------------------------------------
-namespace HealthGateway.Swagger
+namespace HealthGateway.Common.Swagger
 {
     using Microsoft.AspNetCore.Mvc;
     using Microsoft.Extensions.DependencyInjection;
