+++ Apps/Common/Swagger/SwaggerConfiguration.cs
@@ -1,0 +1,58 @@
+﻿//-------------------------------------------------------------------------
+// Copyright © 2019 Province of British Columbia
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//-------------------------------------------------------------------------
+namespace HealthGateway.Common.Swagger
+{
+    using System;
+    using Microsoft.AspNetCore.Builder;
+    using Microsoft.Extensions.Configuration;
+    using Microsoft.Extensions.DependencyInjection;
+
+    /// <summary>
+    /// The swagger configuration class.
+    /// </summary>
+    public static class SwaggerConfiguration
+    {
+        /// <summary>
+        /// Configures the swagger services.
+        /// </summary>
+        /// <param name="services">The service collection provider.</param>
+        /// <param name="config">The configuration provider.</param>
+        public static void ConfigureServices(IServiceCollection services, IConfiguration config)
+        {
+            if (config is null)
+            {
+                throw new ArgumentNullException(nameof(config));
+            }
+
+            services.Configure<SwaggerSettings>(config.GetSection(nameof(SwaggerSettings)));
+
+            services
+                .AddApiVersionWithExplorer()
+                .AddSwaggerOptions()
+                .AddSwaggerGen();
+        }
+
+        /// <summary>
+        /// Configures the app to use swagger.
+        /// </summary>
+        /// <param name="app">The application builder provider.</param>
+        public static void Configure(IApplicationBuilder app)
+        {
+            // Enable middleware to serve swagger-ui (HTML, JS, CSS, etc.),
+            app.UseSwaggerDocuments();
+        }
+    }
+}
