+++ Apps/Common/common.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.28307.329
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "common", "common.csproj", "{8660083F-F582-47C6-BC98-6DCF0F4AC851}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{8660083F-F582-47C6-BC98-6DCF0F4AC851}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8660083F-F582-47C6-BC98-6DCF0F4AC851}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8660083F-F582-47C6-BC98-6DCF0F4AC851}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8660083F-F582-47C6-BC98-6DCF0F4AC851}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {B3A24862-C4B4-4450-AEDF-869107BE8C49}
+	EndGlobalSection
+EndGlobal
