+++ Apps/Immunization/src/Immunization.csproj
@@ -35,4 +35,7 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="4.0.1" />    
     <PackageReference Include="Newtonsoft.Json" Version="12.0.2" />
   </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\..\Common\Common.csproj" />
+  </ItemGroup>
 </Project>
