+++ Apps/Patient/Patient.sln
@@ -1,7 +1,10 @@
-
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Patient", "src\Patient.csproj", "{3DEB1A43-6D33-4F3F-95D6-91BA2F613185}"
+VisualStudioVersion = 15.0.28307.329
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Patient", "src\Patient.csproj", "{3DEB1A43-6D33-4F3F-95D6-91BA2F613185}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Common", "..\Common\Common.csproj", "{8BE4E9AE-164C-495A-A9B5-A07135053998}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -13,5 +16,15 @@ Global
 		{3DEB1A43-6D33-4F3F-95D6-91BA2F613185}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3DEB1A43-6D33-4F3F-95D6-91BA2F613185}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3DEB1A43-6D33-4F3F-95D6-91BA2F613185}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8BE4E9AE-164C-495A-A9B5-A07135053998}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8BE4E9AE-164C-495A-A9B5-A07135053998}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8BE4E9AE-164C-495A-A9B5-A07135053998}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8BE4E9AE-164C-495A-A9B5-A07135053998}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {A5B0AD0C-27D5-42E0-8D0D-472FEF5BC637}
 	EndGlobalSection
 EndGlobal
