--- Apps/Patient/src/Swagger/SwaggerDefaultValues.cs
@@ -1,63 +1,0 @@
-﻿//-------------------------------------------------------------------------
-// Copyright © 2019 Province of British Columbia
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//-------------------------------------------------------------------------
-namespace HealthGateway.Swagger
-{
-    using System.Linq;
-    using Swashbuckle.AspNetCore.Swagger;
-    using Swashbuckle.AspNetCore.SwaggerGen;
-
-    /// <summary>
-    /// Represents the Swagger/Swashbuckle operation filter used to document the implicit API version parameter.
-    /// </summary>
-    /// <remarks>This <see cref="IOperationFilter"/> is only required due to bugs in the <see cref="SwaggerGenerator"/>.
-    /// Once they are fixed and published, this class can be removed.</remarks>
-    public sealed class SwaggerDefaultValues : IOperationFilter
-    {
-        /// <summary>
-        /// Applies the filter to the specified operation using the given context.
-        /// </summary>
-        /// <param name="operation">The operation to apply the filter to.</param>
-        /// <param name="context">The current operation filter context.</param>
-        public void Apply(Operation operation, OperationFilterContext context)
-        {
-            if (operation != null && context != null && operation.Parameters != null)
-            {
-                foreach (var parameter in operation.Parameters.OfType<NonBodyParameter>())
-                {
-                    var description = context.ApiDescription.ParameterDescriptions.First(p => p.Name == parameter.Name);
-                    var routeInfo = description.RouteInfo;
-
-                    if (parameter.Description == null)
-                    {
-                        parameter.Description = description.ModelMetadata?.Description;
-                    }
-
-                    if (routeInfo == null)
-                    {
-                        continue;
-                    }
-
-                    if (parameter.Default == null)
-                    {
-                        parameter.Default = routeInfo.DefaultValue;
-                    }
-
-                    parameter.Required |= !routeInfo.IsOptional;
-                }
-            }
-        }
-    }
-}
