+++ Apps/WebClient/WebClient.sln
@@ -3,12 +3,14 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.26124.0
 MinimumVisualStudioVersion = 15.0.26124.0
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "test", "testUnit\test.csproj", "{DE237219-CFAC-4186-B909-4A8F3312ABC4}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "test", "testUnit\test.csproj", "{DE237219-CFAC-4186-B909-4A8F3312ABC4}"
 	ProjectSection(ProjectDependencies) = postProject
 		{4EF62115-3012-4911-B7AA-3B798231ADC4} = {4EF62115-3012-4911-B7AA-3B798231ADC4}
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebClient", "src\WebClient.csproj", "{4EF62115-3012-4911-B7AA-3B798231ADC4}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebClient", "src\WebClient.csproj", "{4EF62115-3012-4911-B7AA-3B798231ADC4}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Common", "..\Common\Common.csproj", "{95231CE5-4C0B-45E4-B485-632FE24A3950}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -19,22 +21,7 @@ Global
 		Release|x64 = Release|x64
 		Release|x86 = Release|x86
 	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{4EF62115-3012-4911-B7AA-3B798231ADC4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4EF62115-3012-4911-B7AA-3B798231ADC4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4EF62115-3012-4911-B7AA-3B798231ADC4}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{4EF62115-3012-4911-B7AA-3B798231ADC4}.Debug|x64.Build.0 = Debug|Any CPU
-		{4EF62115-3012-4911-B7AA-3B798231ADC4}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{4EF62115-3012-4911-B7AA-3B798231ADC4}.Debug|x86.Build.0 = Debug|Any CPU
-		{4EF62115-3012-4911-B7AA-3B798231ADC4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4EF62115-3012-4911-B7AA-3B798231ADC4}.Release|Any CPU.Build.0 = Release|Any CPU
-		{4EF62115-3012-4911-B7AA-3B798231ADC4}.Release|x64.ActiveCfg = Release|Any CPU
-		{4EF62115-3012-4911-B7AA-3B798231ADC4}.Release|x64.Build.0 = Release|Any CPU
-		{4EF62115-3012-4911-B7AA-3B798231ADC4}.Release|x86.ActiveCfg = Release|Any CPU
-		{4EF62115-3012-4911-B7AA-3B798231ADC4}.Release|x86.Build.0 = Release|Any CPU
 		{DE237219-CFAC-4186-B909-4A8F3312ABC4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{DE237219-CFAC-4186-B909-4A8F3312ABC4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DE237219-CFAC-4186-B909-4A8F3312ABC4}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -47,5 +34,35 @@ Global
 		{DE237219-CFAC-4186-B909-4A8F3312ABC4}.Release|x64.Build.0 = Release|Any CPU
 		{DE237219-CFAC-4186-B909-4A8F3312ABC4}.Release|x86.ActiveCfg = Release|Any CPU
 		{DE237219-CFAC-4186-B909-4A8F3312ABC4}.Release|x86.Build.0 = Release|Any CPU
+		{4EF62115-3012-4911-B7AA-3B798231ADC4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4EF62115-3012-4911-B7AA-3B798231ADC4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4EF62115-3012-4911-B7AA-3B798231ADC4}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{4EF62115-3012-4911-B7AA-3B798231ADC4}.Debug|x64.Build.0 = Debug|Any CPU
+		{4EF62115-3012-4911-B7AA-3B798231ADC4}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{4EF62115-3012-4911-B7AA-3B798231ADC4}.Debug|x86.Build.0 = Debug|Any CPU
+		{4EF62115-3012-4911-B7AA-3B798231ADC4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4EF62115-3012-4911-B7AA-3B798231ADC4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4EF62115-3012-4911-B7AA-3B798231ADC4}.Release|x64.ActiveCfg = Release|Any CPU
+		{4EF62115-3012-4911-B7AA-3B798231ADC4}.Release|x64.Build.0 = Release|Any CPU
+		{4EF62115-3012-4911-B7AA-3B798231ADC4}.Release|x86.ActiveCfg = Release|Any CPU
+		{4EF62115-3012-4911-B7AA-3B798231ADC4}.Release|x86.Build.0 = Release|Any CPU
+		{95231CE5-4C0B-45E4-B485-632FE24A3950}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{95231CE5-4C0B-45E4-B485-632FE24A3950}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{95231CE5-4C0B-45E4-B485-632FE24A3950}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{95231CE5-4C0B-45E4-B485-632FE24A3950}.Debug|x64.Build.0 = Debug|Any CPU
+		{95231CE5-4C0B-45E4-B485-632FE24A3950}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{95231CE5-4C0B-45E4-B485-632FE24A3950}.Debug|x86.Build.0 = Debug|Any CPU
+		{95231CE5-4C0B-45E4-B485-632FE24A3950}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{95231CE5-4C0B-45E4-B485-632FE24A3950}.Release|Any CPU.Build.0 = Release|Any CPU
+		{95231CE5-4C0B-45E4-B485-632FE24A3950}.Release|x64.ActiveCfg = Release|Any CPU
+		{95231CE5-4C0B-45E4-B485-632FE24A3950}.Release|x64.Build.0 = Release|Any CPU
+		{95231CE5-4C0B-45E4-B485-632FE24A3950}.Release|x86.ActiveCfg = Release|Any CPU
+		{95231CE5-4C0B-45E4-B485-632FE24A3950}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {80946C82-3F9C-4F96-B46F-59A297C44E36}
 	EndGlobalSection
 EndGlobal
