--- Apps/WebClient/src/Server/Swagger/ServicesExtensions.cs
@@ -1,64 +1,0 @@
-﻿//-------------------------------------------------------------------------
-// Copyright © 2019 Province of British Columbia
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//-------------------------------------------------------------------------
-namespace HealthGateway.WebClient.Swagger
-{
-    using Microsoft.AspNetCore.Mvc;
-    using Microsoft.Extensions.DependencyInjection;
-    using Microsoft.Extensions.Options;
-    using Swashbuckle.AspNetCore.Swagger;
-    using Swashbuckle.AspNetCore.SwaggerGen;
-    using Swashbuckle.AspNetCore.SwaggerUI;
-
-    /// <summary>
-    /// Service Collection(IServiceCollection) Extensions.
-    /// </summary>
-    public static class ServicesExtensions
-    {
-        /// <summary>
-        /// Add AddVersionedApiExplorer and AddApiVersioning middlewares.
-        /// </summary>
-        /// <param name="services">services.</param>
-        /// <returns>IServiceCollection.</returns>
-        public static IServiceCollection AddApiVersionWithExplorer(this IServiceCollection services)
-        {
-            return services
-                .AddVersionedApiExplorer(options =>
-                {
-                    options.GroupNameFormat = "'v'VVV";
-                    options.SubstituteApiVersionInUrl = true;
-                })
-                .AddApiVersioning(options =>
-                {
-                    options.AssumeDefaultVersionWhenUnspecified = true;
-                    options.ReportApiVersions = true;
-                    options.DefaultApiVersion = new ApiVersion(1, 0);
-                });
-        }
-
-        /// <summary>
-        /// Add swagger services.
-        /// </summary>
-        /// <param name="services"><see cref="IServiceCollection"/>/>.</param>
-        /// <returns>IServiceCollection.</returns>
-        public static IServiceCollection AddSwaggerOptions(this IServiceCollection services)
-        {
-            return services
-                .AddTransient<IConfigureOptions<SwaggerOptions>, ConfigureSwaggerOptions>()
-                .AddTransient<IConfigureOptions<SwaggerUIOptions>, ConfigureSwaggerUiOptions>()
-                .AddTransient<IConfigureOptions<SwaggerGenOptions>, ConfigureSwaggerGenOptions>();
-        }
-    }
-}
